Implement device groups for manual triggering
The --host option now supports groups of switches defined in the
config.json under "devgroups" key.  This is only supported with --trigger
option being used in conjuction with --host.  The syntax is
--host=@group, where 'group' is group name.

This enables manual check-in on group of hosts with a single
cvs-logwatcher.pl execution, ie. it's no longer needed to litter crontab
with entry for each host that requires manual check-in.