Revise organization of the configuration
This commit changes the structure of the configuration by splitting the former
"logfiles" section into "logfiles" and "targets".

The "logfiles" section now only defines logfiles and their match regexes.

The "targets" section is where the actions to perform are defined.
"targets" section is now list of individual targets.  Upon match in a
logfile, the targets are searched for a match and the first matching
target is then executed.  Currently there's no new functionality, so the
only matching criteria is "logfile", but more can be added in the future.

This commit is designed to enable one logfile to be served by multiple
targets (for example devices that are handled using SNMP, devices handled
by SSH session etc.)