Implement "validrange" feature for targets
This feature allows filter out of junk from the start and end of a config
file saved using an expect log (ie.  connection messages, echo of issued
commands etc.) The "validrange" key is a two-element array with starting
and ending regexp that constrains the parts of the config that are
considered the configuration itself.  Only this part of config is
commited into the repository.  Either "validrange" element can be left
null to not constrain start/end of the valid range of the configuration.