New "filter" option for targets
This allows filtering configs by sets of regular expressions and is a
complement to the "validrange" option.  All lines matching any of the
regexes in "filter" are removed from the configuration.  Like
"validrange" this can be used to get rid of text that is not actually
configuration (artifacts of the scraping process etc.)