Change fs layout and how development mode is triggered
Previous to this commit, the filesystem was fixed in place by absolute
prefix path hardcoded in the main script.  The development mode was
triggered when the program found, that the last element of the path is
'dev'.

This commit changes this mechanic. There is no longer the hardcoded
pathname within the main script and the main script must be executed from
the base directory (that is, as bin/cvs-logwatcher.pl).  Additionally,
two command-line options were implemented: --debug and --devel.

--debug raises debug level for the logging to DEBUG

--devel invokes --debug and switches logging to STDOUT

The logging appender switching requires, that the Log4perl logger now
configures both appenders AScrn and AFile.  The program will then remove
either one depending on the commandline.