Separate tftp directory from temp directory
config.tftproot+config.tftpdir were used as temporary directory for
retrieving configurations even when these were received via other means
(SSH, telnet).  This commits separates the two usages.

Also, this restores the original way of specifying tftproot and tftpdir
separately.  The tftpdir (subdir in tftproot) is needed because that is
what is sent to the device, so we need to have that separate.