Remove getting hostname from SNMP
This gets rid of all SNMP interaction in the code (apart from the Cisco
writeNet relict which will be also removed).  The idea was to get host's
own idea of its hostname, not whatever is in DNS or gets logged in
syslog.  Unfortunately, this adds significant complexity -- and we
haven't even supported SNMPv3.  The hostname can be parsed from the
configuration file anyway, so there is no reason to query the host over
SNMP.