Enable multifile processing
One task now can produce multiple files.  Note that this is fairly
preliminary and won't work with --nocheckin=FILE and possibly have other
issues.  Also all the processing options will be applied to all files
uniformly, which limits usefulness to just variants of configurations --
the intent of this is to be able to save both standard and full
configuration (configuration with defaults listed).