Implement target configuration option 'files'
Up until now we were getting configuration files from by recording the
Expect sessions and issuing some kind of 'show configuration' command on
the device.  But this doesn't work when we actually get the configuration
via scp/sftp or similar fetching method, as there's nothing to record.
To facilitate such way of getting configuration, it is now possible to
specify one or more file using the 'files' key under 'targets' entry.
This file is checked for existence after the Expect session finishes and
is added to the list of files to be processed.  Replacement tokens and
directories are allowed in the file names.