Implement ability to stash data for later use
Sometimes, all data needed to craft commit are not available in one log
entry and it is useful to be able to store some information parsed from a
log entry for later use.  This commit implements two elements that make this
possible, the third part of this infrastructure was added in previous
commit.  The first one is Stash singleton, that is able to store information
partitioned by hostname.  The second element is the optional 'action'
property of 'target' configuration, which specifies a callback that gets a
reference to the Stash instance and reference to a hash of named capture
groups from regex match on log entry.  This callback can move things from
the capture groups to the stash for later user.