Separate configuration of regexes from logs
This commit changes the organization of the configuration file.
Previously regular expressions used to match against log entries were
configured in the "logfiles" section.  This had the issue if one wanted
to use the same regex with multiple logfiles, it was necessary to include
the exact same regex multiple times, which violates the DRY principle.

To improve, we now relegate regular expressions into their own config
section, "matches", where they are enumerated for later use.  The
"matches" section is a hash and has the form of MATCHID => REGEX.

The reduced "logfiles" section now omits the regex defintion and single
entry looks like this: LOGID => [ FILE, MATCHID1, MATCHID2, ..., MATCHIDN
], where MATCHIDn refers to regular expression ids defined in the
"matches" section.

This commit also disables unit tests for the Config module, these need to
be completely revamped in the future.